+++ test-e2e/step-definitions/index.ts
@@ -1,5 +1,6 @@
 import { Then } from '@cucumber/cucumber';
 import { expect } from 'chai';
+import { type MemoryValue } from '@qavajs/core';
 
 Then('I expect {string} memory value to be equal {string}', async function(actual, expected) {
     const actualValue = await this.memory.getValue(actual);
@@ -16,4 +17,8 @@ Then('I expect viewport size to equal {string}', async function (expectedSize) {
     const expectedValue = await this.memory.getValue(expectedSize);
     const actualValue = await this.wdio.browser.getWindowSize();
     expect(actualValue).to.deep.equal(expectedValue, 'Viewport size do not match');
-})
+});
+
+Then('I save {value} to memory as {value}', async function (value: MemoryValue, key: MemoryValue) {
+     key.set(await value.value());
+});