+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 :pencil: - chore
 
 ## [2.0.0]
+- :pencil: added memory processor to playwrightLocator parameter type
 - :rocket: new page object approach
 
 ## [0.55.1]